allow concurrent sending of >1 MB frames
When sending large frames, they get split up into chunks of 1 MB each.
When this happens, the individual chunks may get interleaved with other
`send` requests, potentially corrupting the data of the large frame.
To prevent this, a queue is introduced that holds the individual `send`
requests to be processed in order.