Don't sleep 1ms after every send (#25)
The loop in doSend is intended to pause 1ms in between 1MB chunks of
a very large frame. But it pauses even on small frames. This
introduces a 1ms delay after every send, effectively throttling
connections to no more than 1000 messages/second.

I simply changed the logic to sleep 1ms before every chunk *after the
first*. This change literally doubled the throughput of a client I'm
working on.

Fixes #24